# perf: Improve performance of build_style method