```
ruby 3.4.0dev (2024-02-27T20:49:04Z master edc7b73fc4) +YJIT [arm64-darwin23]
Warming up --------------------------------------
                 New    77.023k i/100ms
                 Old    54.003k i/100ms
Calculating -------------------------------------
                 New    827.226k (± 0.6%) i/s -      4.159M in   5.028123s
                 Old    551.346k (± 0.6%) i/s -      2.808M in   5.093489s

Comparison:
                 New:   827226.3 i/s
                 Old:   551346.0 i/s - 1.50x  slower
```